Manual sync from SCM git to GitHub
David Ormrod Morley: Merge pull request 'Add github ci workflow for PLAMS SCMSUITE-9942 SO107' (!27) from DavidOrmrodMorley/SCMSUITE-9942.plams_github_ci_workflow into trunk SO--
David Ormrod Morley: Add python test versions 3.9 and 3.10 and trigger for new version tags for plams ci workflow SCMSUITE-9942 SO107
David Ormrod Morley: Fix PLAMS logger to avoid creating workdir SCMSUITE-8687 SO107
Matti Hellström: fix plams Molecule docstring SO--
Matti Hellström: rearrange/tidy plams docs, move examples and AMS to more prominent positions SO--
Matti Hellström: fix parameter interface not being used for singlepoint, allow None to 'unset' option in plams SCMJob (similar to AMSJob) SO103
David Ormrod Morley: Fix ci workflow too avoid continuing on black error SCMSUITE-9942 SO107
David Ormrod Morley: Merge pull request 'Update PLAMS examples and tests SCMSUITE-9919 SO--' (!24) from DavidOrmrodMorley/SCMSUITE-9919.plams_examples_and_tests into trunk SO--
Robert Rüger: fix the obvious noqa TODOs SCMSUITE-9942 SO--
David Ormrod Morley: Add github ci workflow for PLAMS SCMSUITE-9942 SO107
David Ormrod Morley: Add review todos for flake8 issues in PLAMS SCMSUITE-9942 SO--
David Ormrod Morley: Fix trivial flake8 issues in PLAMS SCMSUITE-9942 SO--
David Ormrod Morley: Fix cross-project links in PLAMS docs SCMSUITE-9919 SO--
David Ormrod Morley: Minor fixes for PLAMS test examples SCMSUITE-9919 SO--
David Ormrod Morley: Remove a couple more init/finish from PLAMS examples SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS examples with small fixes SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example IRSpectrumFromMD SCMSUITE-9919 SO--
David Ormrod Morley: Convert PLAMS ipynb examples subsection headings to subsubsection SCMSUITE-9919 SO--
David Ormrod Morley: Format PLAMS examples with black SCMSUITE-9919 SO--
David Ormrod Morley: Run black formatter for PLAMS generate_example script SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example numgrad  SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example ADFNBO SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example ReorganizationEnergy SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example ADFFrag SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example XRD SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS misc examples SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example AMSPlumedMD SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example BandStructure SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS misc examples SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS examples AMSCalculator SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example PropertyPrediction SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example ReactionsDiscovery SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example ConformersMultipleMolecules SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example ConformersGeneration SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example TuningRangeSeparation SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example ChargeTransferIntegralsADF SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example AMSTSWorkflow SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example ExcitationsWorkflow SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example BasisSetBenchmark SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS examples for MD trajectory analysis SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example PlotCorrelation SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS example MoleculesTable SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example MoleculeSubstitution SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example ManyJobsInParallel SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example He2DissociationCurve SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example AMSSettingsSystem SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS test and example WaterOptimization SCMSUITE-9919 SO--
Paul Spiering: Merge pull request 'Allow for dipoleMoments to propagate in the hybrid engine. SO--' (!15) from hybrid_dipoles into trunk SO--
Matti Hellström: do not interpret imaxmo/MAXMOV as convergence.step reaxff-to-ams conversion SO--
David Ormrod Morley: Merge pull request 'Add template for PLAMS examples SCMSUITE-9919 SO--' (!19) from DavidOrmrodMorley/SCMSUITE-9919.plams_examples into trunk SO--
David Ormrod Morley: Run black formatter on PLAMS SCMSUITE-8687 SO107
David Ormrod Morley: Update PLAMS unit tests to skip when ams installation unavailable SCMSUITE-8687 SO107
David Ormrod Morley: Do not call clean for uninitialised jobmanager in PLAMS finish SCMSUITE-8687 SO107
Matti Hellström: use basename to extract example name in plams generate_examples (allows e.g. trailing slash in directory name) SO107
David Ormrod Morley: Update PLAMS example template readme SCMSUITE-9919 SO--
David Ormrod Morley: Update PLAMS water optimisation example SCMSUITE-9919 SO--
David Ormrod Morley: Update generate_example script to use PLAMS template files SCMSUTIE-9919 SO--
David Ormrod Morley: Add PLAMS example template files SCMSUTIE-9919 SO--
David Ormrod Morley: Convert generate_example script from bash to sh and update readme SCMSUTIE-9919 SO--
David Ormrod Morley: Add readme for PLAMS examples SCMSUTIE-9919 SO--
David Ormrod Morley: Update PLAMS test to use example water optimization notebook SCMSUITE-9919 SO--
David Ormrod Morley: Add script in PLAMS to generate doc examples from notebook and run for water optimization SCMSUITE-9919 SO--
David Ormrod Morley: Update build plams doc script SCMSUITE-9919 SO---
David Ormrod Morley: Add _explicit_init to config settings in PLAMS SCMSUITE-8687 SO107
Robert Rüger: add forgotten quotes around not yet imported JobManager class in finish() typehint SCMSUITE-8687 SO107
Robert Rüger: remove technical note from docs which is no longer true (was not true for a while already: plams importing is now just 'normal') SO--
Robert Rüger: add type hint for plams.finish() function SCMSUITE-8687 SO107
Robert Rüger: fix small type hinting issues SCMSUITE-8687 SO107
David Ormrod Morley: Update docs on PLAMS init and finish functions SCMSUITE-8687 SO--
David Ormrod Morley: Make PLAMS init and finish functions optional SCMSUITE-8687 SO--
David Ormrod Morley: Fix paths for windows in PLAMS function unit test #3 SCMSUITE-8687 SO--
David Ormrod Morley: Fix paths for windows in PLAMS function unit test #2 SCMSUITE-8687 SO--
David Ormrod Morley: Fix paths for windows in PLAMS function unit test SCMSUITE-8687 SO--
David Ormrod Morley: Reinstate support for PLAMS defaults file in code and add unit tests SCMSUITE-8687 SO107
David Ormrod Morley: Update PLAMS defaults docs SCMSUITE-8687 SO107
David Ormrod Morley: Remove usage of plams_defaults file in PLAMS SCMSUITE-8687 SO107
David Ormrod Morley: Remove plams_defaults file SCMSUITE-8687 SO107
Paul Spiering: Allow for dipoleMoments to propagate in the hybrid engine. SO--
David Ormrod Morley:  Add unit tests for PLAMS kf tools SCMSUITE-8674 SO--
David Ormrod Morley: Add unit tests for PLAMS geometry tools SCMSUITE-8674 SO--
David Ormrod Morley: Add PLAMS test rkf file SCMSUITE-8674 SO--
David Ormrod Morley: Add unit tests for PLAMS Bond SCMSUITE-8674 SO--
David Ormrod Morley: Add unit tests for PLAMS Atom and minor property refactoring SCMSUITE-8674 SO--

+++ core/functions.py
@@ -8,11 +8,15 @@ import time
 import types
 from os.path import dirname, expandvars, isdir, isfile
 from os.path import join as opj
-from typing import Dict
+from typing import Dict, Iterable, Optional, TYPE_CHECKING
+import atexit
 
-from scm.plams.core.errors import FileError, PlamsError
+from scm.plams.core.errors import FileError
 from scm.plams.core.private import retry
 from scm.plams.core.settings import Settings, ConfigSettings
+
+if TYPE_CHECKING:
+    from scm.plams.core.jobmanager import JobManager
 
 __all__ = [
     "init",
@@ -28,59 +32,55 @@ __all__ = [
     "read_all_molecules_in_xyz_file",
 ]
 
+# Initialise config with default values for standard settings
+# An initial call to '_init' and a final call to '_finish' are made later in this module
 config = ConfigSettings()
-config.init = False
+# ===========================================================================
+
+_stdlock = threading.Lock()
+_filelock = threading.Lock()
+
+
+def log(message: str, level: int = 0) -> None:
+    """Log *message* with verbosity *level*.
+
+    Logs are printed independently to the text logfile (a file called ``logfile`` in the main working folder) and to the standard output. If *level* is equal or lower than verbosity (defined by ``config.log.file`` or ``config.log.stdout``) the message is printed. Date and/or time can be added based on ``config.log.date`` and ``config.log.time``. All logging activity is thread safe.
+    """
+    if config.init and "log" in config:
+        if level <= config.log.file or level <= config.log.stdout:
+            message = str(message)
+            prefix = ""
+            if config.log.date:
+                prefix += "%d.%m|"
+            if config.log.time:
+                prefix += "%H:%M:%S"
+            if prefix:
+                prefix = "[" + prefix.rstrip("|") + "] "
+                message = time.strftime(prefix) + message
+            if level <= config.log.stdout:
+                with _stdlock:
+                    print(message)
+            if level <= config.log.file and config["default_jobmanager"] is not None:
+                try:
+                    with _filelock, open(config.default_jobmanager.logfile, "a") as f:
+                        f.write(message + "\n")
+                except FileNotFoundError:
+                    pass
+    elif level <= 3:
+        # log() is called before plams.init() was called ...
+        with _stdlock:
+            print(message)
+
+
 # ===========================================================================
 
 
-def init(path=None, folder=None, config_settings: Dict = None, quiet=False, use_existing_folder=False):
-    """Initialize PLAMS environment. Create global ``config`` and the default |JobManager|.
-
-    An empty |Settings| instance is created and populated with default settings by executing ``plams_defaults``. The following locations are used to search for the defaults file, in order of precedence:
-
-    *   If ``$PLAMSDEFAULTS`` variable is in your environment and it points to a file, this file is used (executed as a Python script).
-    *   If ``$AMSHOME`` variable is in your environment and ``$AMSHOME/scripting/scm/plams/plams_defaults`` exists, it is used.
-    *   Otherwise, the path ``../plams_defaults`` relative to the current file (``functions.py``) is checked. If defaults file is not found there, an exception is raised.
-
-    Then a |JobManager| instance is created as ``config.default_jobmanager`` using *path* and *folder* to determine the main working folder. Settings for this instance are taken from ``config.jobmanager``. If *path* is not supplied, the current directory is used. If *folder* is not supplied, ``plams_workdir`` is used.  If *use_existing_folder* is True and the working folder already exists, PLAMS will not create a new working folder with an incremental suffix (e.g. plams_workdir.002). Instead, it will just use the pre-existing folder (note: that this might lead to issues if the working folder is not empty).
-
-    Optionally, an additional `dict` (or |Settings| instance) can be provided to the `config_settings` argument which will be used to update the values from the ``plams_defaults``.
-
-    .. warning::
-      This function **must** be called before any other PLAMS command can be executed. Trying to do anything without it results in a crash. See also |master-script|.
+def _init() -> None:
     """
-
-    if config.init:
-        return
-
-    # Update all standard options to their defaults
-    config.update(ConfigSettings())
-    if "PLAMSDEFAULTS" in os.environ and isfile(expandvars("$PLAMSDEFAULTS")):
-        defaults = expandvars("$PLAMSDEFAULTS")
-    elif "AMSHOME" in os.environ and isfile(opj(expandvars("$AMSHOME"), "scripting", "scm", "plams", "plams_defaults")):
-        defaults = opj(expandvars("$AMSHOME"), "scripting", "scm", "plams", "plams_defaults")
-    else:
-        defaults = opj(dirname(dirname(__file__)), "plams_defaults")
-        if not isfile(defaults):
-            raise PlamsError("plams_defaults not found, please set PLAMSDEFAULTS or AMSHOME in your environment")
-    with open(defaults, "r") as f:
-        exec(compile(f.read(), defaults, "exec"))
-
-    config.update(config_settings or {})
-
-    from scm.plams.core.jobmanager import JobManager
-
-    config.default_jobmanager = JobManager(config.jobmanager, path, folder, use_existing_folder)
-
-    if not quiet:
-        log("Running PLAMS located in {}".format(dirname(dirname(__file__))), 5)
-        log("Using Python {}.{}.{} located in {}".format(*sys.version_info[:3], sys.executable), 5)
-        log("PLAMS defaults were loaded from {}".format(defaults), 5)
-
-        log("PLAMS environment initialized", 5)
-        log("PLAMS working folder: {}".format(config.default_jobmanager.workdir), 1)
-
-    config.slurm = _init_slurm() if "SLURM_JOB_ID" in os.environ else None
+    Internal initialisation of the config settings, which should only be called once.
+    This finishes setup of the config settings, adding any settings from a plams_defaults file, any slurm settings, and marking the settings as initialised.
+    It also attempts to import the dill package and logs a warning if this was unsuccessful.
+    """
 
     try:
         import dill  # noqa F401
@@ -90,11 +90,41 @@ def init(path=None, folder=None, config_settings: Dict = None, quiet=False, use_
             1,
         )
 
+    _load_defaults_file()
+
+    config.slurm = _init_slurm() if "SLURM_JOB_ID" in os.environ else None
+
     config.init = True
 
 
-def _init_slurm():
-    """If PLAMS is running under Slurm, query some information about the batch system and set up the environment for the PLAMS/Slurm integration."""
+def _load_defaults_file() -> Optional[str]:
+    """
+    Searches for a raw python defaults file and executes it to apply settings to the global config.
+    The defaults file is specified either by the ``$PLAMSDEFAULTS`` environment variable,
+    the path ``$AMSHOME/scripting/scm/plams/plams_defaults``, or the path ``../plams_defaults`` relative to the current file.
+
+    Note that this is a legacy functionality which is no longer encouraged but still used.
+
+    :return: path of the defaults file used to populate the global config
+    """
+    if "PLAMSDEFAULTS" in os.environ and isfile(expandvars("$PLAMSDEFAULTS")):
+        defaults = expandvars("$PLAMSDEFAULTS")
+    elif "AMSHOME" in os.environ and isfile(opj(expandvars("$AMSHOME"), "scripting", "scm", "plams", "plams_defaults")):
+        defaults = opj(expandvars("$AMSHOME"), "scripting", "scm", "plams", "plams_defaults")
+    else:
+        defaults = opj(dirname(dirname(__file__)), "plams_defaults")
+    if isfile(defaults):
+        with open(defaults, "r") as f:
+            exec(compile(f.read(), defaults, "exec"))
+        return defaults
+    else:
+        return None
+
+
+def _init_slurm() -> Optional[Settings]:
+    """
+    If PLAMS is running under Slurm, query some information about the batch system and set up the environment for the PLAMS/Slurm integration
+    """
     ret = Settings()
     if "SLURM_JOB_ID" not in os.environ:
         log("Slurm setup aborted: SLURM_JOB_ID is not set")
@@ -146,16 +176,80 @@ def _init_slurm():
     return ret
 
 
+# Initial call to finish set up of config
+_init()
+
+
+def init(
+    path: Optional[str] = None,
+    folder: Optional[str] = None,
+    config_settings: Optional[Dict] = None,
+    quiet: bool = False,
+    use_existing_folder: bool = False,
+) -> None:
+    """
+    Initialize PLAMS environment. Set up the global ``config`` and its default |JobManager|.
+
+    This function only needs to be called to override default values on the global ``config``, without interacting with it directly.
+    For example, both of the following approaches are equivalent:
+
+    .. code-block:: python
+
+        # Call to init
+        init(path="my/path", folder="my_folder", config_settings=my_settings)
+
+        # Direct modification of config
+        config.update(my_settings) # direct modification of config
+        config.default_jobmanager = JobManager(config.jobmanager, "my/path", "my_folder")
+
+    .. note::
+        A call to ``init()`` with no arguments is therefore a no-op, and is not required. It will however force initialisation of the lazy default |JobManager|.
+        This is in contrast to AMS2024 and earlier versions, where this function had to be called before any other PLAMS command could be executed.
+
+    :param path: for the default |JobManager|
+    :param folder: for the default |JobManager|
+    :param use_existing_folder: for the default |JobManager|
+    :param config_settings: |Settings| to update config with - these will overwrite any existing items
+    :param quiet: do not log header with information about the PLAMS environment
+    """
+
+    if config.init and config._explicit_init:
+        return
+
+    # Build the config settings via:
+    # * (Re)Initialise config with default values
+    # * Update with any values from a PLAMS defaults file
+    # * Update with any settings passed to the method
+    config.update(ConfigSettings())
+    defaults_file = _load_defaults_file()
+    config.update(config_settings or {})
+
+    from scm.plams.core.jobmanager import JobManager
+
+    config.default_jobmanager = JobManager(config.jobmanager, path, folder, use_existing_folder)
+
+    config.slurm = _init_slurm() if "SLURM_JOB_ID" in os.environ else None
+
+    if not quiet:
+        log("Running PLAMS located in {}".format(dirname(dirname(__file__))), 5)
+        log("Using Python {}.{}.{} located in {}".format(*sys.version_info[:3], sys.executable), 5)
+        if defaults_file is not None:
+            log("PLAMS defaults were loaded from {}".format(defaults_file), 5)
+        log("PLAMS environment initialized", 5)
+        log("PLAMS working folder: {}".format(config.default_jobmanager.workdir), 1)
+
+    config.init = True
+    config._explicit_init = True
+
+
 # ===========================================================================
 
 
-def finish(otherJM=None):
-    """Wait for all threads to finish and clean the environment.
-
-    This function must be called at the end of your script for |cleaning| to take place. See |master-script| for details.
-
-    If you used some other job managers than just the default one, they need to be passed as *otherJM* list.
+def _finish():
     """
+    Internal clean up of the PLAMS environment, which will be called at the end of the script.
+    """
+
     if not config.init:
         return
 
@@ -163,17 +257,45 @@ def finish(otherJM=None):
         if thread.name == "plamsthread":
             thread.join()
 
-    config.default_jobmanager._clean()
+    # Only clean the default lazy job manager if it has been initialised
+    # as otherwise accessing it will create an empty workdir
+    if config["default_jobmanager"] is not None:
+        config.default_jobmanager._clean()
+
+        if config.erase_workdir is True:
+            shutil.rmtree(config.default_jobmanager.workdir)
+
+    config.init = False
+
+
+def finish(otherJM: Optional[Iterable["JobManager"]] = None):
+    """
+    Clean up the PLAMS environment. This can be explicitly called for |cleaning| to take place.
+    If you used some other job managers than just the default one, they need to be passed as *otherJM*.
+
+    A call is registered to be performed at exit, which will perform cleaning for the default job manager.
+
+    .. note::
+        A call to ``finish()`` with no arguments once at the end of a script is therefore a no-op, and is not required.
+        This is in contrast to AMS2024 and earlier versions, where this function had to be called at the end of a script for |cleaning| to take place.
+
+    :param otherJM: additional job managers used in a workflow to be cleaned
+    """
+    if not config.init:
+        return
+
+    _finish()
+
     if otherJM:
         for jm in otherJM:
             jm._clean()
+
     log("PLAMS environment cleaned up successfully", 5)
     log("PLAMS run finished. Goodbye", 3)
 
-    if config.erase_workdir is True:
-        shutil.rmtree(config.default_jobmanager.workdir)
 
-    config.init = False
+# Register call to _finish on workflow end
+atexit.register(_finish)
 
 
 # ===========================================================================
@@ -280,44 +402,6 @@ def read_all_molecules_in_xyz_file(filename):
             except FileError:
                 break
     return mols
-
-
-# ===========================================================================
-
-
-_stdlock = threading.Lock()
-_filelock = threading.Lock()
-
-
-def log(message, level=0):
-    """Log *message* with verbosity *level*.
-
-    Logs are printed independently to the text logfile (a file called ``logfile`` in the main working folder) and to the standard output. If *level* is equal or lower than verbosity (defined by ``config.log.file`` or ``config.log.stdout``) the message is printed. Date and/or time can be added based on ``config.log.date`` and ``config.log.time``. All logging activity is thread safe.
-    """
-    if config.init and "log" in config:
-        if level <= config.log.file or level <= config.log.stdout:
-            message = str(message)
-            prefix = ""
-            if config.log.date:
-                prefix += "%d.%m|"
-            if config.log.time:
-                prefix += "%H:%M:%S"
-            if prefix:
-                prefix = "[" + prefix.rstrip("|") + "] "
-                message = time.strftime(prefix) + message
-            if level <= config.log.stdout:
-                with _stdlock:
-                    print(message)
-            if level <= config.log.file and "default_jobmanager" in config:
-                try:
-                    with _filelock, open(config.default_jobmanager.logfile, "a") as f:
-                        f.write(message + "\n")
-                except FileNotFoundError:
-                    pass
-    elif level <= 3:
-        # log() is called before plams.init() was called ...
-        with _stdlock:
-            print(message)
 
 
 # ===========================================================================
